# Conflicts:
#	com.vmturbo.action.orchestrator/src/main/java/com/vmturbo/action/orchestrator/action/ActionModeCalculator.java
#	com.vmturbo.action.orchestrator/src/main/java/com/vmturbo/action/orchestrator/store/LiveActionStore.java
#	com.vmturbo.action.orchestrator/src/test/java/com/vmturbo/action/orchestrator/action/ActionDescriptionBuilderTest.java
#	com.vmturbo.components.common/src/main/java/com/vmturbo/components/common/setting/EntitySettingSpecs.java
#	com.vmturbo.market.component/src/main/java/com/vmturbo/market/topology/conversions/TopologyConverter.java
#	com.vmturbo.topology.processor/src/main/java/com/vmturbo/topology/processor/group/settings/EntitySettingsApplicator.java

+++ com.vmturbo.topology.processor/src/test/java/com/vmturbo/topology/processor/history/percentile/PercentileCommodityDataTest.java
@@ -73,7 +73,8 @@ public class PercentileCommodityDataTest extends BaseGraphRelatedTest {
     public void testInit() throws HistoryCalculationException {
         float cap = 100f;
         int used = 70;
-        TopologyEntity entity = mockEntity(1, 1, commType, cap, used, null, null, null, null, true);
+        TopologyEntity entity = mockEntity(1, 1, commType, cap, (double)used, null, null, null,
+            null, true);
         ICommodityFieldAccessor accessor = Mockito.spy(new CommodityFieldAccessor(
                         mockGraph(Collections.singleton(entity))));
         Mockito.doReturn((double)cap).when(accessor).getCapacity(field);
